+++ milkman-auth/src/main/java/milkman/plugin/auth/oauth/Oauth2KeyEditor.java
@@ -30,6 +30,7 @@ public class Oauth2KeyEditor implements KeyEditor<Oauth2Credentials>, ToasterAwa
     private final GenericBinding<Oauth2Credentials, String> clientSecretBinding = GenericBinding.of(Oauth2Credentials::getClientSecret, Oauth2Credentials::setClientSecret);
     private final GenericBinding<Oauth2Credentials, String> scopesBinding = GenericBinding.of(Oauth2Credentials::getScopes, Oauth2Credentials::setScopes);
     private final GenericBinding<Oauth2Credentials, Boolean> autoRefreshBinding = GenericBinding.of(Oauth2Credentials::isAutoRefresh, Oauth2Credentials::setAutoRefresh);
+    private final GenericBinding<Oauth2Credentials, Boolean> autoIssueBinding = GenericBinding.of(Oauth2Credentials::isAutoIssue, Oauth2Credentials::setAutoIssue);
     private final GenericBinding<Oauth2Credentials, Boolean> requestBodyAuthSchemeBinding = GenericBinding.of(Oauth2Credentials::isRequestBodyAuthScheme, Oauth2Credentials::setRequestBodyAuthScheme);
 
     private Toaster toaster;
@@ -54,14 +55,21 @@ public class Oauth2KeyEditor implements KeyEditor<Oauth2Credentials>, ToasterAwa
         root.add(formEntry("Client Id", clientIdBinding, keyEntry));
         root.add(formEntry("Client Secret", clientSecretBinding, keyEntry));
         root.add(formEntry("Scopes", scopesBinding, keyEntry));
+
         var autoRefresh = new JFXToggleButton();
         autoRefresh.setText("Refresh Token on expiry");
         autoRefreshBinding.bindTo(autoRefresh.selectedProperty(), keyEntry);
+
+        var autoIssue = new JFXToggleButton();
+        autoIssue.setText("Issue new token if necessary");
+        autoIssueBinding.bindTo(autoIssue.selectedProperty(), keyEntry);
+
+
         var requestBodyAuthScheme = new JFXToggleButton();
         requestBodyAuthScheme.setText("Credentials in Body");
         requestBodyAuthSchemeBinding.bindTo(requestBodyAuthScheme.selectedProperty(), keyEntry);
 
-        root.add(new HBox(autoRefresh, requestBodyAuthScheme));
+        root.add(new HBox(autoRefresh, autoIssue, requestBodyAuthScheme));
 
         var combobox = root.add(new JFXComboBox<GrantTypeBuilder>());
         combobox.getItems().addAll(