+++ milkman-auth/src/main/java/milkman/plugin/auth/oauth/model/Oauth2Credentials.java
@@ -18,10 +18,14 @@ public class Oauth2Credentials extends KeyEntry {
     String accessTokenEndpoint;
     String scopes;
     boolean autoRefresh;
+    boolean autoIssue;
     boolean requestBodyAuthScheme;
 
     @JsonIgnore
     boolean refreshFailed;
+
+    @JsonIgnore
+    boolean autoIssueFailed;
 
     Oauth2Grant grantType;
     OAuth2Token token;
@@ -53,6 +57,16 @@ public class Oauth2Credentials extends KeyEntry {
                 refreshFailed = true;
             }
         }
+        //if still expired, refresh didnt work
+        if (isExpired() && isAutoIssue() && !autoIssueFailed) {
+            try {
+                fetchNewToken();
+            } catch (Exception e) {
+                log.error("Failed to issue token", e.getMessage());
+                autoIssueFailed = true;
+            }
+        }
+
         return token.getAccessToken();
     }